requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
openpyxl==3.1.2
